% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_lasso.R
\name{gauss}
\alias{gauss}
\title{Gauss function}
\source{
https://github.com/Zongrui-Dai/Kernel-lasso-feature-expansion
}
\usage{
gauss(d1, d2, sigma = 0.5)
}
\arguments{
\item{d1}{vector1}

\item{d2}{vector2}

\item{sigma}{The hyperparameter of RBF kernel function, which indicates the width.}
}
\value{
Calculate the Gauss function
}
\description{
Gauss function
}
\examples{
##
data(iris,package = 'datasets')
w<-gauss(iris[,1],iris[,2])
print(w)
}
\author{
Zongrui Dai
}
\keyword{Gauss}
\keyword{function}
